Refactor reminder service to handle multiple types
Removed `VaccinationRemindersForTomorrowService` and replaced it with `RoundsRemindersForTomorrowService`, which now manages both vaccination and health check reminders. Introduced an enum `ReminderType` for differentiation. Added email templates for health check reminders and created `healthcheck-reminder-template.html`. Updated service registration in `ApplicationServiceExtensions.cs` to reflect these changes.
